Update C99 datatypes

+++ src/batteryloggerdialog.cpp
@@ -26,28 +26,14 @@ void BatteryLoggerDialog::exec() {
 	gmenu2x->setBacklight(100);
 	gmenu2x->setCPU(gmenu2x->confInt["cpuMax"]);
 
-// // skinConfColor
-// 	int i = 0, j = 0;
-// 	for(ConfColorHash::iterator curr = skinConfColor.begin(); curr != skinConfColor.end(); curr++) {
-// 		i++;
-// 		if (i > 31) {
-// 			j++;
-// 			i=0;
-// 		}
-// 		if (j > 14) break;
-
-// 		DEBUG("COLOR: %d,%d %s = %d", i,j, curr->first.c_str(),  (unsigned short)curr->second.r);
-// 		this->bg->box(2+i*10,2+j*10,8,8, curr->second);
-// 	}
-
 	gmenu2x->s->flip();
 
 	MessageBox mb(gmenu2x, gmenu2x->tr["Welcome to the Battery Logger.\nMake sure the battery is fully charged.\nAfter pressing OK, leave the device ON until\nthe battery has been fully discharged.\nThe log will be saved in 'battery.csv'."]);
 	mb.exec();
 
-	uint firstRow = 0, rowsPerPage = gmenu2x->listRect.h/gmenu2x->font->getHeight();
+	uint32_t firstRow = 0, rowsPerPage = gmenu2x->listRect.h/gmenu2x->font->getHeight();
 
-	long tickNow = 0, tickStart = SDL_GetTicks(), tickBatteryLogger = -1000000;
+	int32_t tickNow = 0, tickStart = SDL_GetTicks(), tickBatteryLogger = -1000000;
 	string logfile = gmenu2x->getExePath()+"battery.csv";
 
 	char buf[100];
@@ -85,7 +71,7 @@ void BatteryLoggerDialog::exec() {
 
 		this->bg->blit(gmenu2x->s,0,0);
 
-		for (uint i = firstRow; i < firstRow + rowsPerPage && i < log.size(); i++) {
+		for (uint32_t i = firstRow; i < firstRow + rowsPerPage && i < log.size(); i++) {
 			int rowY, j = log.size() - i - 1;
 			if (log.at(j)=="----") { //draw a line
 				rowY = 42 + (int)((i - firstRow + 0.5) * gmenu2x->font->getHeight());
